third_party/blink/web_tests/external/wpt/app-history/app-history-entry/key-id-location-reload-transitionWhile.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => { window.onload = resolve; });


  const original = i.contentWindow.appHistory.current;
  const { key, id } = original;

  i.contentWindow.appHistory.addEventListener("navigate", e => e.transitionWhile(Promise.resolve()));

  i.onload = t.unreached_func("the iframe must not reload");

  await i.contentWindow.location.reload();

  assert_equals(i.contentWindow.appHistory.current, original);
  assert_equals(i.contentWindow.appHistory.current.key, key);
  assert_equals(i.contentWindow.appHistory.current.id, id);
}, "AppHistoryEntry's key and id after location.reload() intercepted by transitionWhile()");
</script>
